refactor : 참가자 추상 클래스 도입
