+++ src/main/java/blackjack/domain/Player.java
@@ -1,39 +1,23 @@
 package blackjack.domain;
 
-public class Player {
+public class Player extends Participant {
 
-    private final String name;
-    private final Hand hand = new Hand();
 
     public Player(String name) {
-        this.name = name;
+        super(name,new Hand());
+
     }
 
-
+@Override
     public void recieveCard(Card card){
         if(!hand.isBust()){
             hand.addCard(card);
         }
     }
 
-    public int getCardCount(){
-        return hand.getCount();
-    }
-
     public String getName() {
-        return name;
+        return super.getName();
     }
 
-    public String getCardNames(){
-        return hand.getCardNames();
-    }
-
-    public boolean isBust() {
-        return hand.isBust();
-    }
-
-    public int getTotalPoint(){
-        return hand.getTotalPoint();
-    }
 
 }
